fix: ensure accurate total sleep time and sleep score calculation by excluding awake phases in DailySleepData

+++ Wfit-sleep/app/src/main/java/com/example/wfit/presentation/components/SleepCycleGraph.kt
@@ -75,7 +75,7 @@ fun SleepCycleGraph(
             val totalMinutes = ChronoUnit.MINUTES.between(
                 sleepData.cycles.first().startTime,
                 sleepData.cycles.last().endTime
-            )
+            ).toFloat()
 
             var currentX = 0f
             sleepData.cycles.forEach { cycle ->
